separate files for workerasgname resource
