+++ service/controller/v24/resource/workerasgname/delete.go
@@ -1,0 +1,9 @@
+package workerasgname
+
+import (
+	"context"
+)
+
+func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}
